Extract vector and matrix functions into their own files
